FINALLY got the parrying system complete using coroutines, as well as getting the damage that should be applied to the character as a result of the parry quality

+++ Assets/Resources/EntityStats.cs
@@ -5,6 +5,9 @@ public class EntityStats : ScriptableObject
 {
     public int maxHp;
     public int attack;
-    public int parryWindow;
+    public double parryWindow;
     public float parryCooldown;
+
 }
+
+
